+++ BambooTracker/io/raw_adpcm_io.cpp
@@ -1,0 +1,50 @@
+/*
+ * Copyright (C) 2022 Rerrah
+ *
+ * Permission is hereby granted, free of charge, to any person
+ * obtaining a copy of this software and associated documentation
+ * files (the "Software"), to deal in the Software without
+ * restriction, including without limitation the rights to use,
+ * copy, modify, merge, publish, distribute, sublicense, and/or sell
+ * copies of the Software, and to permit persons to whom the
+ * Software is furnished to do so, subject to the following
+ * conditions:
+ *
+ * The above copyright notice and this permission notice shall be
+ * included in all copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+ * EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+ * OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+ * HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY,
+ * WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+ * FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
+ * OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+#include "raw_adpcm_io.hpp"
+#include "file_io_error.hpp"
+#include "io_utils.hpp"
+
+namespace io
+{
+RawAdpcmIO::RawAdpcmIO() : AbstractInstrumentIO("spb", "ADPCM sample", true, false) {}
+
+AbstractInstrument* RawAdpcmIO::load(const BinaryContainer& ctr, const std::string& fileName,
+								 std::weak_ptr<InstrumentsManager> instMan, int instNum) const
+{
+	std::shared_ptr<InstrumentsManager> instManLocked = instMan.lock();
+	int sampIdx = instManLocked->findFirstAssignableSampleADPCM();
+	if (sampIdx < 0) throw FileCorruptionError(FileType::Inst, 0);
+
+	auto adpcm = new InstrumentADPCM(instNum, fileName, instManLocked.get());
+	adpcm->setSampleNumber(sampIdx);
+
+	instManLocked->storeSampleADPCMRawSample(sampIdx, ctr.toVector());
+	instManLocked->setSampleADPCMRootKeyNumber(sampIdx, 60);	// o5c
+	instManLocked->setSampleADPCMRootDeltaN(sampIdx, 0x49cd);	// 16000Hz
+
+	return adpcm;
+}
+}